Workaround for Issue 16 * vc-clearcase.el (clearcase-checkout-comment-type): change default to 'none (vc-clearcase-checkout): signal an error if `clearcase-checkout-comment-type' is not 'none (clearcase-bind-checkout-comment): don't add it to `vc-before-checkin-hook'
Update issue 16

I disabled `clearcase-checkout-comment-type' for now, to make the package
usable.  It won't insert the checkout comment into the checkin buffer and as a
precaution it won't allow the user to specify a checkout comment when file is
checked out from Emacs.  Unfortunately, the user can still specify a checkout
comment form outside, and this comment will be lost on checkin.

The final solution requires adding `clearcase-bind-checkout-comment' to
`log-edit-hook' and using vc-parent-buffer to get to the original file's fprop